Arrumando bugs, usando range
- Erros da adaptação pra Vector
- Inimigos só atiram quando estão perto da nave (deixar ainda melhor)
- TODO

+++ source/Enemy.h
@@ -7,7 +7,7 @@ typedef struct {
 	Vector pos;
 
 	int health;			/* Vida do inimigo. */
-	int range;			/* Alcance da arma. */
+	double range;			/* Alcance da arma. */
 	double precision;	/* Precisao do inimigo. */
 
 	double gunDir[3];
@@ -23,7 +23,7 @@ typedef struct {
 extern int Enemy_DefaultSize;
 
 /* Constructor para Enemy. */
-Enemy* Enemy_new(int x, int y, int z, double precision, int freq, int range);
+Enemy* Enemy_new(int x, int y, int z, double precision, int freq, double range);
 
 /* Update para uma instancia de Enemy. */
 void Enemy_update(Enemy* this, double dt);
